+++ logic/trade_logic.py
@@ -1,6 +1,6 @@
 from logic.api.trade_api import check_steam_login_api, create_steam_topup_order_api, checking_exchange_rate_api
 
-async def create_steam_topup_order(customer_id, login: str, amount: int):
+async def create_steam_topup_order(customer_id, login: str, amount: float):
     if amount < 0.1:
         return 'Слишком маленькая сумма!'
     resp = await check_steam_login_api(login, amount)
